+++ backends/simplego/dotgeneral_large.go
@@ -249,7 +249,7 @@ func dgCopyOutputBlockToFlat[T interface {
 }
 
 func init() {
-	dotGeneralOutputBlockToFlatDTypeMap.Register(dtypes.BFloat16, dgCopyOutputBlockToFlatBFloat16)
+	dotGeneralOutputBlockToFlatDTypeMap.Register(dtypes.BFloat16, priorityTyped, dgCopyOutputBlockToFlatBFloat16)
 }
 
 // dgCopyOutputBlockToFlatBFloat16 copies the blocked output to a flat output, removing the padding.
@@ -588,7 +588,7 @@ func buildDotGeneralKernel[T PODNumericConstraints](lhs, rhs, output *Buffer, bl
 }
 
 func init() {
-	dotGeneralKernelDTypeMap.Register(dtypes.BFloat16, buildDotGeneralKernelBFloat16)
+	dotGeneralKernelDTypeMap.Register(dtypes.BFloat16, priorityTyped, buildDotGeneralKernelBFloat16)
 }
 
 // buildDotGeneralKernelBFloat16 returns a kernel function that does a DotGeneral (matrix multiplication) of the lhs/rhs block