+++ backends/simplego/exec_special_ops.go
@@ -16,32 +16,32 @@ import (
 )
 
 func init() {
-	nodeExecutors[backends.OpTypeIdentity] = execIdentity
-	nodeExecutors[backends.OpTypeWhere] = execWhere
-	nodeExecutors[backends.OpTypeReshape] = execReshape
-	nodeExecutors[backends.OpTypeTranspose] = execTranspose
-	nodeExecutors[backends.OpTypeBroadcast] = execBroadcast
-	nodeExecutors[backends.OpTypeBroadcastInDim] = execBroadcastInDim
-	nodeExecutors[backends.OpTypeReduceMax] = execReduce
-	nodeExecutors[backends.OpTypeReduceMin] = execReduce
-	nodeExecutors[backends.OpTypeReduceSum] = execReduce
-	nodeExecutors[backends.OpTypeReduceProduct] = execReduce
-	nodeExecutors[backends.OpTypeReduceBitwiseAnd] = execReduce
-	nodeExecutors[backends.OpTypeReduceBitwiseOr] = execReduce
-	nodeExecutors[backends.OpTypeReduceBitwiseXor] = execReduce
-	nodeExecutors[backends.OpTypeReduceLogicalAnd] = execReduce
-	nodeExecutors[backends.OpTypeReduceLogicalOr] = execReduce
-	nodeExecutors[backends.OpTypeReduceLogicalXor] = execReduce
-	nodeExecutors[backends.OpTypeIota] = execIota
-	nodeExecutors[backends.OpTypeGather] = execGather
-	nodeExecutors[backends.OpTypeConcatenate] = execConcatenate
-	nodeExecutors[backends.OpTypeConvertDType] = execConvertDType
-	nodeExecutors[backends.OpTypeScatterMax] = execScatter
-	nodeExecutors[backends.OpTypeScatterMin] = execScatter
-	nodeExecutors[backends.OpTypeScatterSum] = execScatter
-	nodeExecutors[backends.OpTypeSlice] = execSlice
-	nodeExecutors[backends.OpTypeArgMinMax] = execArgMinMax
-	nodeExecutors[backends.OpTypeReduceWindow] = execReduceWindow
+	setNodeExecutor(backends.OpTypeIdentity, priorityGeneric, execIdentity)
+	setNodeExecutor(backends.OpTypeWhere, priorityGeneric, execWhere)
+	setNodeExecutor(backends.OpTypeReshape, priorityGeneric, execReshape)
+	setNodeExecutor(backends.OpTypeTranspose, priorityGeneric, execTranspose)
+	setNodeExecutor(backends.OpTypeBroadcast, priorityGeneric, execBroadcast)
+	setNodeExecutor(backends.OpTypeBroadcastInDim, priorityGeneric, execBroadcastInDim)
+	setNodeExecutor(backends.OpTypeReduceMax, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceMin, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceSum, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceProduct, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceBitwiseAnd, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceBitwiseOr, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceBitwiseXor, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceLogicalAnd, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceLogicalOr, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeReduceLogicalXor, priorityGeneric, execReduce)
+	setNodeExecutor(backends.OpTypeIota, priorityGeneric, execIota)
+	setNodeExecutor(backends.OpTypeGather, priorityGeneric, execGather)
+	setNodeExecutor(backends.OpTypeConcatenate, priorityGeneric, execConcatenate)
+	setNodeExecutor(backends.OpTypeConvertDType, priorityGeneric, execConvertDType)
+	setNodeExecutor(backends.OpTypeScatterMax, priorityGeneric, execScatter)
+	setNodeExecutor(backends.OpTypeScatterMin, priorityGeneric, execScatter)
+	setNodeExecutor(backends.OpTypeScatterSum, priorityGeneric, execScatter)
+	setNodeExecutor(backends.OpTypeSlice, priorityGeneric, execSlice)
+	setNodeExecutor(backends.OpTypeArgMinMax, priorityGeneric, execArgMinMax)
+	setNodeExecutor(backends.OpTypeReduceWindow, priorityGeneric, execReduceWindow)
 
 	// For nodes with multiple outputs:
 	multiOutputsNodeExecutors[backends.OpTypeRNGBitGenerator] = execRNGBitGenerator
@@ -293,7 +293,7 @@ func execReduceMaxGeneric[T PODNumericConstraints](operand, output *Buffer, it *
 	}
 }
 
-func init() { reduceMaxDTypeMap.Register(dtypes.BFloat16, execReduceMaxBFloat16) }
+func init() { reduceMaxDTypeMap.Register(dtypes.BFloat16, priorityTyped, execReduceMaxBFloat16) }
 
 // execReduceMaxBFloat16: use reduceMaxDTypeMa to call it.
 func execReduceMaxBFloat16(operand, output *Buffer, it *reduceOutputIterator, dtype dtypes.DType) {
@@ -330,7 +330,7 @@ func execReduceMinGeneric[T PODNumericConstraints](operand, output *Buffer, it *
 	}
 }
 
-func init() { reduceMinDTypeMap.Register(dtypes.BFloat16, execReduceMinBFloat16) }
+func init() { reduceMinDTypeMap.Register(dtypes.BFloat16, priorityTyped, execReduceMinBFloat16) }
 
 func execReduceMinBFloat16(operand, output *Buffer, it *reduceOutputIterator, dtype dtypes.DType) {
 	// Initialize with the highest value.
@@ -365,7 +365,7 @@ func execReduceSumGeneric[T PODNumericConstraints](operand, output *Buffer, it *
 	}
 }
 
-func init() { reduceSumDTypeMap.Register(dtypes.BFloat16, execReduceSumBFloat16) }
+func init() { reduceSumDTypeMap.Register(dtypes.BFloat16, priorityTyped, execReduceSumBFloat16) }
 
 func execReduceSumBFloat16(operand, output *Buffer, it *reduceOutputIterator, _ dtypes.DType) {
 	// Initialize with 0.
@@ -400,7 +400,9 @@ func execReduceProductGeneric[T PODNumericConstraints](operand, output *Buffer, 
 	}
 }
 
-func init() { reduceProductDTypeMap.Register(dtypes.BFloat16, execReduceProductBFloat16) }
+func init() {
+	reduceProductDTypeMap.Register(dtypes.BFloat16, priorityTyped, execReduceProductBFloat16)
+}
 
 func execReduceProductBFloat16(operand, output *Buffer, it *reduceOutputIterator, _ dtypes.DType) {
 	// Initialize with 1.
@@ -719,7 +721,7 @@ func execIotaGeneric[T PODNumericConstraints](params ...any) any {
 	return nil
 }
 
-func init() { dispatchIota.Register(dtypes.BFloat16, execIotaBFloat16) }
+func init() { dispatchIota.Register(dtypes.BFloat16, priorityTyped, execIotaBFloat16) }
 
 func execIotaBFloat16(params ...any) any {
 	output, batchSize, iotaSize, repeatsSize := params[0].(*Buffer), params[1].(int), params[2].(int), params[3].(int)
@@ -1145,8 +1147,8 @@ func execConvertDTypeToBool[FromT PODNumericConstraints, _ bool](operand, output
 
 func init() {
 	// Manually register bool x bfloat16 conversion functions.
-	convertDTypePairMap.Register(dtypes.BFloat16, dtypes.Bool, execConvertDTypeBFloat16ToBool)
-	convertDTypePairMap.Register(dtypes.Bool, dtypes.BFloat16, execConvertDTypeBoolToBFloat16)
+	convertDTypePairMap.Register(dtypes.BFloat16, dtypes.Bool, priorityTyped, execConvertDTypeBFloat16ToBool)
+	convertDTypePairMap.Register(dtypes.Bool, dtypes.BFloat16, priorityTyped, execConvertDTypeBoolToBFloat16)
 }
 
 func execConvertDTypeBFloat16ToBool(operand, output *Buffer) {
@@ -1385,9 +1387,9 @@ var (
 )
 
 func init() {
-	combineMaxDTypeMap.Register(dtypes.BFloat16, combineForScatterMaxBFloat16)
-	combineMinDTypeMap.Register(dtypes.BFloat16, combineForScatterMinBFloat16)
-	combineSumDTypeMap.Register(dtypes.BFloat16, combineForScatterSumBFloat16)
+	combineMaxDTypeMap.Register(dtypes.BFloat16, priorityTyped, combineForScatterMaxBFloat16)
+	combineMinDTypeMap.Register(dtypes.BFloat16, priorityTyped, combineForScatterMinBFloat16)
+	combineSumDTypeMap.Register(dtypes.BFloat16, priorityTyped, combineForScatterSumBFloat16)
 }
 
 func combineForScatterMaxGeneric[T PODNumericConstraints](a, b T) T {
@@ -1647,7 +1649,7 @@ func execArgMinMaxGeneric[T PODNumericConstraints](
 }
 
 func init() {
-	argMinMaxDTypeMap.Register(dtypes.BFloat16, execArgMinMaxGenericBFloat16)
+	argMinMaxDTypeMap.Register(dtypes.BFloat16, priorityTyped, execArgMinMaxGenericBFloat16)
 }
 
 func execArgMinMaxGenericBFloat16(
@@ -1848,10 +1850,10 @@ var (
 )
 
 func init() {
-	reduceWindowMaxDTypeMap.Register(dtypes.BFloat16, reduceWindowMaxBuildUpdateFnBFloat16)
-	reduceWindowMinDTypeMap.Register(dtypes.BFloat16, reduceWindowMinBuildUpdateFnBFloat16)
-	reduceWindowSumDTypeMap.Register(dtypes.BFloat16, reduceWindowSumBuildUpdateFnBFloat16)
-	reduceWindowProductDTypeMap.Register(dtypes.BFloat16, reduceWindowProductBuildUpdateFnBFloat16)
+	reduceWindowMaxDTypeMap.Register(dtypes.BFloat16, priorityTyped, reduceWindowMaxBuildUpdateFnBFloat16)
+	reduceWindowMinDTypeMap.Register(dtypes.BFloat16, priorityTyped, reduceWindowMinBuildUpdateFnBFloat16)
+	reduceWindowSumDTypeMap.Register(dtypes.BFloat16, priorityTyped, reduceWindowSumBuildUpdateFnBFloat16)
+	reduceWindowProductDTypeMap.Register(dtypes.BFloat16, priorityTyped, reduceWindowProductBuildUpdateFnBFloat16)
 }
 
 // Generic functions that build a function that will update the output at outputFlatIdx from the operand at operandFlatIdx.