+++ internal/cmd/simplego_dispatcher/main.go
@@ -180,7 +180,7 @@ func init() {
 {{- $dispatcher := .Dispatcher }}
 {{- $generic := .Generic }}
 {{- range .DTypes }}
-	{{$dispatcher}}.RegisterIfNotSet(dtypes.{{.DType}}, {{$generic}}[{{.GoType}}])
+	{{$dispatcher}}.Register(dtypes.{{.DType}}, priorityGeneric, {{$generic}}[{{.GoType}}])
 {{- end }}
 {{- end }}
 
@@ -190,7 +190,7 @@ func init() {
 {{- $mapName := .MapName }}
 {{- $generic := .Generic }}
 {{- range .DTypes }}
-	{{$mapName}}.RegisterIfNotSet(dtypes.{{.DType}}, {{$generic}}[{{.GoType}}])
+	{{$mapName}}.Register(dtypes.{{.DType}}, priorityGeneric, {{$generic}}[{{.GoType}}])
 {{- end }}
 {{- end }}
 
@@ -204,7 +204,7 @@ func init() {
 {{- $dtype1 := .DType }}
 {{- $goType1 := .GoType }}
 {{- range $dtypes2 }}
-	{{$mapName}}.RegisterIfNotSet(dtypes.{{$dtype1}}, dtypes.{{.DType}}, {{$generic}}[{{$goType1}}, {{.GoType}}])
+	{{$mapName}}.Register(dtypes.{{$dtype1}}, dtypes.{{.DType}}, priorityGeneric, {{$generic}}[{{$goType1}}, {{.GoType}}])
 {{- end }}
 {{- end }}
 {{- end }}